requirements.txt
requests
djangorestframework-jwt
djangorestframework
asgiref==3.2.10
Django==3.1.2
pytz==2020.4
sqlparse==0.4.1
